requirements.txt
praw==7.7.1
beautifulsoup4==4.12.3
requests==2.31.0
google-generativeai==0.8.3
python-dotenv==1.0.0
pandas==2.1.4
lxml==5.1.0
